greetings

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS user (
-  id bigint(20) NOT NULL AUTO_INCREMENT,
+  id INT (20) NOT NULL AUTO_INCREMENT,
   name varchar(100) NOT NULL,
   email varchar(100) DEFAULT NULL,
   address varchar(100) NOT NULL,
